macb: Send link speed to the PL via GPIOs
Added support for an EMIO GPIO for telling the GMII-RGMII conversion
logic in the PL what the link speed is. This is necessary for 1Gb
Ethernet to work on many NI targets.

These changes also magically fix the secondary Ethernet speed LED.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>
Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Natinst-ReviewBoard-ID: 84563
(joshc: updated to use gpiod_* APIs)
(bm: update to gpiod_* APIs)
Signed-off-by: Josh Cartwright <joshc@ni.com>