nirtfeatures: set HARD_BOOT_N if necessary when the driver loads
On systems where FPGA autoload is controlled by software, we need a way
to determine if a system has just had power applied. This is necessary
to implement the autoload on every power-on feature. Once the FPGA has
been autloaded after power-on, software sets a bit in the CPLD so that
subsequent (non power-on) resets don't cause the FPGA to be autoloaded
again. The CPLD provides the HARD_BOOT_N bit for this purpose.

On systems where FPGA autoload is controlled by hardware, such as Smasher,
we need to set this bit at some point after power-on so that software can
correctly determine the reset state of the controller. Since software has
no insight into the status of FPGA autoload on these systems, we can just
set this bit if necessary when the driver loads.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>
[cvadrevu: Fixed trivial conflicts with
  91d93a7e35e0 ("nirtfeatures: Only register bi-color LEDs on supported targets")]
  6c83dd7f8821 ("nirtfeatures: Added NI RT features driver")]
Signed-off-by: Chaitanya Vadrevu <chaitanya.vadrevu@emerson.com>