i2c: cadence: Support reading the bus number from the device tree
When multiple I2C busses are present, we need a way to deterministically
specify the bus number for each bus. Here we add support for reading a
bus-id field from the device tree. We must use i2c_add_numbered_adapter
when specifying a bus number. We use a default value of -1 if the bus
number isn't specified in the device tree, which gives the same behavior
as using i2c_add_adapter.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>