wl12xx:  Disallow channels above 11 until beacons are seen
In order to prevent us from sending out anything on channels
that are illegal in some countries, we wait until we see
beacons from the AP before trying to connect on those channels.

This change is to retain parity with NI's shipped 3.2 kernel for
purposes of compliance, so it should not be upstreamed.

Signed-off-by: Tim Ousley <tim.ousley@ni.com>
[reworked for 3.14 branch and for coding style ]
Signed-off-by: James Minor <james.minor@ni.com>