ath6kl: turn power on/off even if CONFIG_PM is disabled
For normal operation of MMC devices, the bus will turn off the
cards only when CONFIG_PM is enabled and all funcs are disabled.
Make ath6kl_sdio_power_on() and ath6kl_sdio_power_off() actually
turn the power on/off in this case.

This API is used in the same way by the wl12xx driver, which has
the same issues.

Signed-off-by: James Minor <james.minor@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>

Natinst-ReviewBoard-ID: 130170
Natinst-CAR-ID: 566029