nati_zynq_defconfig: Enable Network Emulator
In efd0798 (nat_x86_64_defconfig: Enable Network Emulator), the
i210 EtherCAT support called for NET_SCH_NETEM to force best-effort
traffic on the ECAT interface to be dropped. In the interest of
keeping parity between the arches, the change is only 100k added,
and the chance that some customers may find this useful, add
the configuration to ARM as well.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>