auxdisplay: oel9m1027: Truly OEL9M1027 OLED module driver
Truly OEL9M1027 OLED display module are not supported at this moment.
This commit is to add driver to support Truly OEL9M1027 OLED module.

This driver will interact with Sino Wealth SH1107 driver and provide
a framebuffer to userspace (/dev/fbX) for user application to interact
with OLED display.

Cc: Miguel Ojeda Sandonis <miguel.ojeda.sandonis@gmail.com>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Wilson Lee <wilson.lee@ni.com>
Acked-by: Julia Cartwright <julia@ni.com>
Natinst-ReviewBoard-ID: 205126
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
[cvadrevu: Fixed trivial conflicts due to newer module]
[cvadrevu: Kconfig: Changed --help-- to help]
Signed-off-by: Chaitanya Vadrevu <chaitanya.vadrevu@emerson.com>