net: ath6kl: silex: Separate region and board file selection
To support the single antenna board files for products with 1
antenna, separate out the selection of the region from the
selection of the board file.

On OF platforms, there will now be 2 device tree entries:
atheros,region-code - The region code (like US)
atheros,board-id - The board file code (like 00, 10, etc)

In the process, do a small refactor of the
CONFIG_ATH6KL_NI_BIOS_DOMAIN case to be more consistent with the
device tree case.

Signed-off-by: James Minor <james.minor@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Nathan Sullivan <nathan.sullivan@ni.com>
Acked-by: Kyle Roeschley <kyle.roeschley@ni.com>
Natinst-ReviewBoard-ID: 223288
Natinst-CAR-ID: 686866