nati_zynq_defconfig: disable CONFIG_LEGACY_PTYS
Legacy PTYs are disabled on most desktop operating systems and are not believed
to be in use on NI Linux RT; they were not in use in Debian as of 2008 [1]. The
default number of legacy PTYs (256) also increases the size of /dev on a
cRIO-9068 by over 4x (!). Disabling them is also believe to slightly improve
boot time. Finally, the KConfig help alludes to "a number of problems, including
security" with legacy PTYs.

Those are all fine reasons to make this change -- but the proximate
justification for this change is a little more subtle: legacy PTYs complicate
heuristics for enumerating serial devices. Nominally, you could reasonably use
something like "any chardev with path /dev/tty[^0-9].*" to include e.g. ttyS and
ttyPS and ttyUSB, but not VCs or /dev/tty. But the existence of legacy PTYs mean
you'd need to match something more like /dev/tty[^0-9a-ep-z].* -- assuming you
had fewer than 256 PTYs, of course.

[1] http://ubuntu.5.x6.nabble.com/Jaunty-config-changes-Reduce-CONFIG-LEGACY-PTY-COUNT-to-0-td366355.html

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>