serial: 8250: add driver for NI UARTs
The National Instruments (NI) 16550 is a 16550-like UART with larger
FIFOs and embedded RS-232/RS-485 transceiver control circuitry. This
patch adds a driver that can operate this UART, which is used for
onboard serial ports in several NI embedded controller designs.

Portions of this driver were originally written by Jaeden Amero and
Karthik Manamcheri, with extensive cleanups and refactors since.

Signed-off-by: Brenda Streiff <brenda.streiff@ni.com>
Cc: Gratian Crisan <gratian.crisan@ni.com>
Cc: Jason Smith <jason.smith@ni.com>
[cvadrevu: fix trivial conflict with 66ebe67d1b68b ("tty/serial: 8250: Sort drivers in Makefile")]
Signed-off-by: Chaitanya Vadrevu <chaitanya.vadrevu@emerson.com>